API Refactor AppKernel into CoreKernel
