+++ main.php
@@ -1,7 +1,7 @@
 <?php
 
 use SilverStripe\Control\HTTPRequestBuilder;
-use SilverStripe\Core\AppKernel;
+use SilverStripe\Core\CoreKernel;
 use SilverStripe\Core\HTTPApplication;
 use SilverStripe\Core\Startup\ErrorControlChainMiddleware;
 
@@ -11,7 +11,7 @@ require __DIR__ . '/src/includes/autoload.php';
 $request = HTTPRequestBuilder::createFromEnvironment();
 
 // Default application
-$kernel = new AppKernel(BASE_PATH);
+$kernel = new CoreKernel(BASE_PATH);
 $app = new HTTPApplication($kernel);
 $app->addMiddleware(new ErrorControlChainMiddleware($app));
 $response = $app->handle($request);
